backend/requirements.txt
fastapi==0.104.1
psycopg2-binary==2.9.9
uvicorn==0.24.0.post1
python-dotenv==1.0.0
pydantic-settings==2.1.0

